Try many things to find problem and implement logger
There is a problem with the resulting state in the plots.
To find the source of the problem I tried a lot of different things and
finally landed on printing everything with a logger.
The logger is working but I have yet to look at the output to find the
problem.